+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- fix: Rate limiting for cached envelope items #685
 - fix: Crash when SentryClient is nil in SentryHub #681
 - feat: Send cached envelopes first #676
 - feat: Use envelopes for sending events #650